COM-3633: refacto table
